+++ NEWS.md
@@ -23,7 +23,8 @@ can be made.
 [memoised](https://github.com/r-lib/memoise) so they are
 only computed once even adding several layers with the same contours 
 (`geom_contour() + geom_text_contour()`) or running the same plot while tweaking
-it's appearance. 
+it's appearance.
+- New `geom_shadow()` for casting shadows in topographic maps.
 
 # metR 0.1
 - New function: `GetTopography()`